Discord Bot/requirements.txt
discord.py
requests
dotenv
scikit-learn
